SpectraFS

Documentation for SpectraFS.

SpectraFS.band_avg

SpectraFS.centeredFFT

SpectraFS.confid

SpectraFS.spectralbasis

SpectraFS.spectralpowerlaw

SpectraFS.totalspectralenergy

SpectraFS.band_avg — Function

function bandavg.jl Block averages for band averaging [yyavg]=band_avg(yy,num,dimension)

Inputs: yy, quantity to be averaged (must be vector or matrix)

num, number of bands to average
dimension (optional), dimension to average along; if specified, must be 1 or 2

Tom Farrar, 2016, jfarrar@whoi.edu Ported to Julia, Jake Gebbie, 2021, jgebbie@whoi.edu =#

SpectraFS.centeredFFT — Method

function centeredFFT(x,dt)

Computes FFT, with zero frequency in the center, and returns dimensional frequency vector.

Inputs: x, vector to be transformed dt, time increment

Adapted from a function written by Quan Quach of blinkdagger.com Tom Farrar, 2016, jfarrar@whoi.edu For Julia, Jake Gebbie, 2021, jgebbie@whoi.edu

SpectraFS.confid — Method

function confid(α,ν)

Help with computing confidence intervals

should be sigma^2/S^2 confidence bounds where sigma^2 is true variance
check value (J&W) is alpha =.05, nu=19, lower bound is .58
upper bound is 2.11

SpectraFS.spectralbasis — Function

function spectralbasis(t,f)

basis function to reconstruct mean ocean temperature (Θ̄)
on the t temporal grid

Arguments

t: times of interest

f: Fourier frequencies

includemean=false::Bool: include the mean value in the basis set?,

Output

A::Matrix: each column is an independent basis function, first (nt-1)/2 columns are sine coefficients second (nt-1)/2 columns are cosine coefficients last column represents the mean value

SpectraFS.spectralpowerlaw — Function

spectralpowerlaw(β,f) = f.^-β

Arguments

f: frequencies

β: power law coefficient, low frequencies

e: total energy

βhi: power law coefficient, high frequencies

Output

Φ: spectral energy density

SpectraFS.totalspectralenergy — Method

function totalspectralenergy(Φ,f)

Arguments

Φ: power spectral density

f: Fourier frequencies

Output

e: total energy